+++ Netable/Example/Views/RootView.swift
@@ -14,29 +14,9 @@ struct RootView: View {
 
     var body: some View {
         VStack {
-            if let error = viewModel.error {
-                VStack {
-                    HStack {
-                        Spacer()
-                        Text("Error: \(error)")
-                        Spacer()
-                        Image(systemName: "xmark")
-                            .foregroundColor(.black)
-                            .onTapGesture {
-                                viewModel.clearError()
-                            }
-                    }
-                }.frame(maxWidth: .infinity)
-                .padding()
-                .background(.yellow)
-                .onAppear {
-                    DispatchQueue.main.asyncAfter(deadline: .now() + 3) {
-                        viewModel.error = nil
-                    }
-                }
-            }
+            ErrorView(viewModel: viewModel)
             if viewModel.user == nil {
-                loginView
+                LoginView(viewModel: viewModel.loginVM)
             } else {
                 TabView {
                     HomeView(viewModel: viewModel.homeVM)
@@ -61,43 +41,11 @@ struct RootView: View {
                 }
             }
         }.onAppear {
-            viewModel.bindViewModel()
-            }
-            .onDisappear {
-                viewModel.unbindViewModel()
-            }
+        viewModel.bindViewModel()
+        }
+        .onDisappear {
+            viewModel.unbindViewModel()
+        }
     }
 
-    var loginView: some View {
-        VStack(alignment: .center, spacing: 10) {
-            Spacer()
-            VStack {
-                Text("Welcome back!")
-                    .font(.title)
-                TextField("Username", text: $viewModel.username)
-                    .padding()
-                    .background(.white)
-                    .cornerRadius(10)
-                    .padding(.bottom, 10)
-                    .textInputAutocapitalization(.never)
-                SecureField("Password", text: $viewModel.password)
-                    .padding()
-                    .background(.white)
-                    .cornerRadius(10)
-                    .padding(.bottom, 10)
-                Button(action: { viewModel.login() }) {
-                    Text("Login")
-                }.padding()
-                    .padding(.horizontal, 40)
-                    .background(.blue)
-                    .cornerRadius(10)
-                    .foregroundColor(.white)
-                Button("Add Credentials") {
-                    viewModel.fillForm()
-                }
-            }.padding()
-            Spacer()
-        }.background(Color.lightGrey)
-
-    }
 }
